feat: manage patient record on chaincode

+++ src/main/java/healthscape/com/healthscape/users/api/UserApi.java
@@ -1,10 +1,12 @@
 package healthscape.com.healthscape.users.api;
 
-import healthscape.com.healthscape.fabric.dto.PatientRecordDto;
+import healthscape.com.healthscape.fabric.dto.MyChaincodePatientRecordDto;
+import healthscape.com.healthscape.fabric.dto.ChaincodePatientRecordDto;
 import healthscape.com.healthscape.fabric.service.FabricUserService;
 import healthscape.com.healthscape.fhir.dtos.FhirUserDto;
 import healthscape.com.healthscape.fhir.service.FhirService;
 import healthscape.com.healthscape.file.service.FileService;
+import healthscape.com.healthscape.patientRecords.service.PatientRecordService;
 import healthscape.com.healthscape.security.model.UserTokenState;
 import healthscape.com.healthscape.security.service.AuthenticationService;
 import healthscape.com.healthscape.shared.ResponseJson;
@@ -24,6 +26,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.io.IOException;
 import java.net.URI;
 import java.util.List;
 
@@ -36,9 +39,10 @@ public class UserApi {
 
     private final UserService userService;
     private final AuthenticationService authenticationService;
-    private final FabricUserService fabricUserService;
     private final UsersMapper usersMapper;
     private final FhirService fhirService;
+    private final FabricUserService fabricUserService;
+    private final PatientRecordService patientRecordService;
     private final FileService fileService;
 
     @GetMapping("")
@@ -49,14 +53,20 @@ public class UserApi {
 
     @PostMapping("")
     public ResponseEntity<?> registerUser(@RequestBody RegisterDto user) {
-        AppUser appUser = userService.register(user, "ROLE_PATIENT");
-        byte[] photo = new byte[0];
+        AppUser appUser = null;
         try {
-            // TODO: uncomment
+            appUser = userService.register(user);
+            ChaincodePatientRecordDto chaincodePatientRecordDto = fhirService.registerPatient(appUser, user.getIdentifier());
             fabricUserService.registerUser(appUser);
-            //            photo = fhirService.registerPatient(appUser, user.getIdentifier());
+            if (!chaincodePatientRecordDto.isExisting()) {
+                patientRecordService.createPatientRecord(appUser, chaincodePatientRecordDto);
+            }else{
+                patientRecordService.updatePatientRecord(appUser, chaincodePatientRecordDto);
+            }
         } catch (Exception e) {
-            userService.deleteUser(appUser);
+            if(appUser != null) {
+                userService.deleteUser(appUser);
+            }
             return ResponseEntity.badRequest().body(new ResponseJson(400, e.getMessage()));
         }
         return ResponseEntity.created(URI.create(ServletUriComponentsBuilder.fromCurrentContextPath().toUriString())).body(usersMapper.userToUserDto(appUser));
@@ -64,12 +74,19 @@ public class UserApi {
 
     @PutMapping(value = "", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
     public ResponseEntity<?> updateUser(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestPart("userDto") FhirUserDto userDto, @RequestPart("image") MultipartFile image) {
+        String imagePath = null;
         try {
-            String imagePath = fileService.saveImageToStorage(image);
+            imagePath = fileService.saveImageToStorage(image);
             userDto.setImagePath(imagePath);
-            userService.updateUser(token, userDto);
-            fhirService.updateUser(token, userDto);
+            AppUser user = userService.getUserFromToken(token);
+            MyChaincodePatientRecordDto myChaincodePatientRecordDto = patientRecordService.getMyPatientRecord(user);
+            MyChaincodePatientRecordDto updatedMyPatientRecordDto = fhirService.updatePatient(userDto, myChaincodePatientRecordDto.getOfflineDataUrl());
+            patientRecordService.updateMyPatientRecord(user, updatedMyPatientRecordDto);
+            userService.updateUser(user, userDto, imagePath);
+        } catch (IOException e) {
+            return ResponseEntity.badRequest().body(new ResponseJson(400, e.getMessage()));
         } catch (Exception e) {
+            fileService.deleteImage(imagePath);
             return ResponseEntity.badRequest().body(new ResponseJson(400, e.getMessage()));
         }
 
@@ -86,7 +103,9 @@ public class UserApi {
         UserTokenState tokens;
         try {
             AppUser user = userService.changeEmail(token, email);
-            fhirService.changeEmail(user.getId().toString(), email);
+            MyChaincodePatientRecordDto myChaincodePatientRecordDto = patientRecordService.getMyPatientRecord(user);
+            MyChaincodePatientRecordDto updatedMyPatientRecordDto = fhirService.changeEmail(myChaincodePatientRecordDto.getOfflineDataUrl(), email);
+            patientRecordService.updateMyPatientRecord(user, updatedMyPatientRecordDto);
             tokens = authenticationService.getAuthentication(user);
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(new ResponseJson(400, e.getMessage()));
